Fix an issue with 32 bit transaction id's

+++ src/VPay.Ols.Processor/Commands/OlsFile/ProcessNonFinancial.cs
@@ -77,7 +77,7 @@ public static class ProcessNonFinancial
 
             originalFile.Details.RemoveAll(d => !OPTUM_BINS.Contains(int.Parse(d.Bin)));
 
-            var tpaResults = await _mediator.Send(new GetClientForTransactions.Query(originalFile.Details.Where(d => !string.IsNullOrWhiteSpace(d.SeExternalIdNumber)).Select(d => int.Parse(d.SeExternalIdNumber)).ToList()), cancellationToken).ConfigureAwait(false);
+            var tpaResults = await _mediator.Send(new GetClientForTransactions.Query(originalFile.Details.Where(d => !string.IsNullOrWhiteSpace(d.SeExternalIdNumber)).Select(d => long.Parse(d.SeExternalIdNumber)).ToList()), cancellationToken).ConfigureAwait(false);
             foreach (var detailRecord in originalFile.Details)
             {
                 detailRecord.CardNumber = $"{detailRecord.CardNumber[..6]}XXXXXX{detailRecord.CardNumber[^4..]}";
@@ -91,7 +91,7 @@ public static class ProcessNonFinancial
 
                 if (!string.IsNullOrWhiteSpace(detailRecord.SeExternalIdNumber))
                 {
-                    var tpaResult = tpaResults.Find(t => t.TransactionId == int.Parse(detailRecord.SeExternalIdNumber));
+                    var tpaResult = tpaResults.Find(t => t.TransactionId == long.Parse(detailRecord.SeExternalIdNumber));
 
                     if (tpaResult == null)
                     {
